requirements.txt
sanic
requests
rsa
cryptography
colorama
pydantic
websocket-client
flask 
rich